Shape state on the backend
This is something about superglue that is unique among solution. And
that's its philosophy on shaping state:

__The backend is responsible for shaping UI state. Contrast this
with traditional approaches that pull data from the server, shape
the data on the client side, then passes them on to components.__

I classify Both REST and GraphQL (slightly mitigated since you are
able to specify some shape) as traditional approaches here. Because
once we pulled the data from the backend, we spend a lot of time
in JS land to shape it to the props React component uses.

But you know, ruby's larger standard library, Rails helpers,
expressiveness, is even BETTER than JS for shaping data.

In this commit I moved some of the presentational aspects
of your dashboard to the backend. I'm shaping my props to roughly
how I view the designs. And I'm awarded for it, my react component
usage is very easy, i just splat the props in.

Lastly, I changed the key format to be camelCase from snake_case.
This is because in JS land, i prefer to work with camelCase, and
since props is where i do transformations, i handle the conversion
there. Its not automatic, that's intentional, being explicit
about the keys will speed up your app since we don't have to run
`camelize` in ruby land.