Add defer :manual
This prevents the block from running at all on initial load. To make it run, you will
have to dig for it. And that's what you have with `props_at=data.search.results`.

Its a perfect usecase for search, but the cool part of this is that the same
pattern `defer: :manual` pattern can be used for modals, tabs, toasts, anything
that is initially hidden but then shown later.

This primitive of digging into a tree to find missing pieces of __your own shaped data__
is one of Superglue's best features.

The other half of `defer: :manual` is `defer: :auto`, which fetches for missing
data automatically. This is essentially a form of Progressive JSON:
https://overreacted.io/progressive-json/

Inertia does not have strong primitives, which is why they can't support
features like modals, and have resorted to having the community build for
missing functionality. https://github.com/inertiaui/modal